refactor: migrate attribution to drift-sdk
(cherry picked from commit 273ff5bf8227a3f86088193e4d719a9971c62255)
